Extract commands.rs tests into commands_tests.rs (under 1000-line limit)
The #171 scrub helpers + regression tests pushed src/runtime/commands.rs to 1028 lines, over the 1000-line hard limit enforced by check-source-file-size.sh. Move the whole inline #[cfg(test)] mod tests into src/runtime/commands_tests.rs via #[path], matching the existing convention used by harness/runner, app_input, state, etc. No test or behavior change.
